solve typos

+++ mpi_openmp/include/kmeans.h
@@ -15,7 +15,7 @@
 
 void elementPaddedIntArray(int *array, int value, int size);
 
-void elementPaddedFloatArray(int *array, int value, int size);
+void elementPaddedFloatArray(float *array, float value, int size);
 
 void elementIntArray(int *array, int value, int size);
